fixup d9f23f2f822a59771fdc3cab648785d4f651e1b2

+++ include/linux/damon.h
@@ -761,7 +761,7 @@ struct damon_filter {
 	bool matching;
 	bool allow;
 	union {
-		u64 memcg_id;
+		unsigned short memcg_id;
 	};
 	struct list_head list;
 };
